+++ .gitignore
@@ -1,7 +1,29 @@
-/.metadata/
-/sistema_poo_patrimonio/.settings/
-/sistema_poo_patrimonio/bin/
-/sistema_poo_patrimonio/.classpath/
-/sistema_poo_patrimonio/.project/
+### IntelliJ IDEA ###
+out/
+!**/src/main/**/out/
+!**/src/test/**/out/
 
+### Eclipse ###
+.apt_generated
+.classpath
+.factorypath
+.project
+.settings
+.springBeans
+.sts4-cache
+bin/
+!**/src/main/**/bin/
+!**/src/test/**/bin/
 
+### NetBeans ###
+/nbproject/private/
+/nbbuild/
+/dist/
+/nbdist/
+/.nb-gradle/
+
+### VS Code ###
+.vscode/
+
+### Mac OS ###
+.DS_Store
