--- sistema_poo_patrimonio/src/models/Relatorio.java
@@ -1,5 +1,0 @@
-package models;
-
-public class Relatorio {
-
-}
