--- sistema_poo_patrimonio/src/module-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * 
- */
-module sistema_poo_patrimonio {
-}
