--- sistema_poo_patrimonio/src/sistema_poo_patrimonio/Main.java
@@ -1,13 +1,0 @@
-package sistema_poo_patrimonio;
-
-import views.HomeView;
-
-public class Main {
-	public static void main(String[] args) {
-		HomeView hm = new HomeView();
-		hm.exibirHomeView();
-		
-		
-	}
-
-}
