+++ src/Main.java
@@ -1,0 +1,8 @@
+import views.HomeView;
+
+public class Main {
+    public static void main(String[] args) {
+        HomeView hm = new HomeView();
+        hm.exibirHomeView();
+    }
+}
