+++ src/models/PatrimonioModel.java
@@ -1,0 +1,4 @@
+package models;
+
+public class PatrimonioModel {
+}
